Add anime.csv file generated from anime_scraper
Added the final output of the anime scraper in the form of a CSV file named "anime.csv". The file contains information about 5530 different anime shows that were scraped from a website Findanime.net. The information includes the title, short title, URL of the show, a source of the image, tags, translation availability, description, number of episodes, and user rating. The CSV file was generated using the script in "anime_scraper.ipynb" Jupyter Notebook. The data in the file is organized and structured, making it easy to use for further analysis or to build an application.